Update datascrubbing.sh
added ability for script to handle the situation where a BTRFS volume has never performed a scrub process before. this can happen if the volume was just created. 
added ability for script to recognize if RAID array does not support scrubbing (raid 0, raid1, or raid10) and act accordingly 
added ability for script to recognize if RAID array is repairing, degraded, or recovering and act accordingly